on_counter_billing_project/requirements.txt
asgiref==3.8.1
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
cloudinary==1.39.1
Django==4.2.11
django-environ==0.11.2
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==21.2.0
importlib_resources==6.4.0
inflection==0.5.1
install==1.3.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mysqlclient==2.2.4
packaging==24.0
phonenumbers==8.13.34
pillow==10.3.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
zipp==3.18.1
